+++ R/get_assessments_be.R
@@ -97,7 +97,7 @@
 get_assessments_be <- function(
   date_range = NULL,
   limit = Inf,
-  download = TRUE,
+  download = FALSE,
   cache_dir = NULL,
   overwrite = FALSE,
   max_file_size_mb = NULL,